+++ src/types/index.ts
@@ -26,3 +26,4 @@ export * from "./eth1";
 export * from "./block";
 export * from "./state";
 export * from "./custom";
+export * from "./validator";
