+++ src/types/validator.ts
@@ -1,4 +1,6 @@
 import {BeaconState, bytes48} from "../types";
+import {Shard, ValidatorIndex} from "./index";
+import {Slot} from "./custom";
 
 export interface ValidatorCtx {
   publicKey: bytes48[];
@@ -8,4 +10,10 @@ export interface ValidatorCtx {
 
 export interface GenesisInfo {
   startTime: number;
+}
+
+export interface CommitteeAssignment {
+  validators: ValidatorIndex[];
+  shard: Shard;
+  slot: Slot;
 }
